requirements.txt
fastapi
uvicorn
python-multipart
psutil
httpx
pillow-heif
pillow
